+++ example-no-express/.eslintrc.js
@@ -1,0 +1,17 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+  root: true,
+  parser: '@typescript-eslint/parser',
+  plugins: [
+    '@typescript-eslint',
+  ],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+  ],
+  rules: {
+    "indent": ["error", 2],
+    "semi": ["error"],
+    "no-explicit-any": 0
+  }
+};