+++ example/jest.config.js
@@ -1,8 +1,0 @@
-// eslint-disable-next-line no-undef
-module.exports = {
-  preset: 'ts-jest',
-  collectCoverage: true,
-  collectCoverageFrom: ['src/**/*.ts', '!**/*.d.ts'],
-  testEnvironment: 'node',
-  testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/'],
-};