+++ lib/application/application.ts
@@ -1,43 +1,17 @@
-import Express from "express";
-
-import { ControllerConfiguration } from "../controller/controller";
-import { RouteConfigurator } from "../route/route";
-import { RequestError } from "../error/request.error";
-import { ConfigurationError } from "../error/configuration.error";
-
+import { ControllerConfiguration, ControllerHandler } from "../controller/controller";
+import { RouteConfiguration } from "../types";
 export class AppConfigurator {
   private _prefix = "";
-  private readonly _express: Express.Application;
   private readonly _controllers: ControllerConfiguration<any>[] = [];
   private readonly logRequests: boolean;
-  private _configured: boolean;
 
   /**
    * Create AppConfigurator
    * @param expressApplication - express application
    * @param logRequests - flag if requests should be logged, true by default
    */
-  constructor(expressApplication: Express.Application, logRequests = true) {
-    this._express = expressApplication;
-    this._configured = false;
+  constructor(logRequests = true) {
     this.logRequests = logRequests;
-  }
-
-  /**
-   * Starts application, register controllers routes in express app
-   * and connect to configured port.
-   * @param printConfiguration - print express application routes enabled by default.
-   */
-  configure(printConfiguration = true): void {
-    if (this._configured) {
-      throw new ConfigurationError("Cannot configure application multiple times");
-    } else {
-      this._configured = true;
-    }
-    this.configureControllers();
-    if (printConfiguration) {
-      this.printExpressConfig();
-    }
   }
 
   /**
@@ -47,6 +21,10 @@ export class AppConfigurator {
   prefix(prefix: string): AppConfigurator {
     this._prefix = prefix;
     return this;
+  }
+
+  getPrefix(): string {
+    return this._prefix;
   }
 
   /**
@@ -67,92 +45,57 @@ export class AppConfigurator {
     return this;
   }
 
-  // PRIVATE
-
-  private printExpressConfig() {
-    console.log("Routes registered by Express server:");
-    this.getExpressRoutesAsStrings().forEach(route => console.log(route));
+  getControllers(): ControllerConfiguration<any>[] {
+    return this._controllers;
   }
 
-  private configureControllers() {
-    this._controllers.forEach(controller => this.configureController(controller));
-  }
-
-  private configureController(controller: ControllerConfiguration, controllersChain?: ControllerConfiguration[]) {
-    if (!controllersChain) {
-      controllersChain = [controller];
-    } else {
-      controllersChain = [...controllersChain, controller];
-    }
-    controller.getControllers().forEach(subController => {
-      this.configureController(subController, controllersChain);
-    });
-    controller.getRoutes().forEach(route => {
-      this.registerRouteInExpress(controllersChain, route);
-    });
-  }
-
-  private registerRouteInExpress(controllersChain: ControllerConfiguration[], route: RouteConfigurator) {
-    const controllersPrefix = controllersChain.reduce(
-      (prefixAcc, controller) => AppConfigurator.getRoutePath(prefixAcc, controller.getPrefix()),
-      this._prefix || ""
+  /**
+   * @description
+   * Build routes from registered controllers.
+   * It will create route configuration for each controller and its sub-controllers.
+   * The route configuration will include the path, method, and handler for each route.
+   * The handler will call the controller's handler and then the route's request handler.
+   * If the controller has sub-controllers, it will recursively build routes for them as well.
+   * @returns Array of route configurations built from registered controllers.
+   */
+  buildRoutes(): RouteConfiguration[] {
+    return this._controllers.reduce<RouteConfiguration[]>(
+      (routes, controller) => routes.concat(this.reduceController(controller, this._prefix)),
+      []
     );
-    const routePath = AppConfigurator.getRoutePath(controllersPrefix, route.getPath());
-
-    if (!route.getMethod()) {
-      throw new ConfigurationError(`Route ${routePath} has no method specified`);
-    }
-
-    this._express[route.getMethod()](`/${routePath}`, this.createApplicationRequestHandler(route, controllersChain));
   }
 
-  private createApplicationRequestHandler(
-    route: RouteConfigurator,
-    controllersChain: ControllerConfiguration[]
-  ): Express.RequestHandler {
-    return async (req: Express.Request, res: Express.Response) => {
-      try {
-        let context: unknown;
+  private reduceController(
+    controller: ControllerConfiguration<any, any>,
+    path?: string,
+    handler?: ControllerHandler
+  ): RouteConfiguration[] {
+    return controller
+      .getRoutes()
+      .map<RouteConfiguration>(route => {
+        const routeConfiguration: RouteConfiguration = {
+          path: AppConfigurator.getRoutePath(path || "", controller.getPrefix(), route.getPath()),
+          method: route.getMethod(),
+          handler: async (res, req, context) => {
+            const currentContext = (await handler?.(res, req, context)) || context;
+            const controllerContext = await controller.getHandler()?.(res, req, currentContext);
 
-        for (const controller of controllersChain) {
-          const newContextValue = await controller.getHandler()?.(req, res, context);
-          context = newContextValue ? newContextValue : context;
-        }
+            return route.getRequestHandler()?.(res, req, controllerContext || currentContext);
+          },
+        };
 
-        const response = await route.getRequestHandler()(req, res, context);
-        if (AppConfigurator.canSendResponse(res)) {
-          if (!res.statusCode) {
-            res.status(200);
-          }
-          res.send(response);
-        }
-      } catch (error) {
-        if (AppConfigurator.canSendResponse(res)) {
-          if (error instanceof RequestError) {
-            res.status(error.httpCode || 500).send(error.response || "Internal error");
-          } else {
-            res.status(500).send("Internal error");
-          }
-        }
-      } finally {
-        this.logRequest(req, res);
-      }
-    };
-  }
-
-  private logRequest(req: Express.Request, res: Express.Response) {
-    if (this.logRequests) {
-      console.log(`Request ${req.method}:${req.path} Response status: ${res.statusCode}`);
-    }
-  }
-
-  private getExpressRoutesAsStrings() {
-    return this._express._router.stack.filter(r => r.route).map(AppConfigurator.expressRouteAsString);
-  }
-  // STATIC
-
-  private static expressRouteAsString(r) {
-    return `${Object.keys(r.route.methods)[0].toUpperCase()}:${r.route?.path}`;
+        return routeConfiguration;
+      })
+      .concat(
+        controller.getControllers().reduce((subRoutes, subController) => {
+          return subRoutes.concat(
+            this.reduceController(subController, controller.getPrefix(), async (res, req, context) => {
+              const currentContext = (await handler?.(res, req, context)) || context;
+              return controller.getHandler()?.(res, req, currentContext);
+            })
+          );
+        }, [] as RouteConfiguration[])
+      );
   }
 
   /**
@@ -163,22 +106,8 @@ export class AppConfigurator {
   private static getRoutePath(...paths: string[]) {
     return paths.filter(path => !!path).join(`/`);
   }
-
-  private static canSendResponse(res: Express.Response) {
-    return !res.writableEnded;
-  }
 }
 
-type ApplicationOptions = {
-  app: Express.Application;
-  logRequests?: boolean;
-};
-
-/**
- * Creates application core
- * @param options.app - Express application used by application
- * @param options.logRequests - log requests, enabled by default
- */
-export function Application(options: ApplicationOptions): AppConfigurator {
-  return new AppConfigurator(options.app, options.logRequests);
+export function Application(): AppConfigurator {
+  return new AppConfigurator();
 }