+++ lib/index.ts
@@ -1,6 +1,8 @@
 /* istanbul ignore file */
 
 export { Application, AppConfigurator } from "./application/application";
-export { Controller, ControllerConfiguration, ControllerHandler } from "./controller/controller";
-export { Route, RouteConfigurator, RouteHandler } from "./route/route";
+export { Controller, ControllerConfiguration, type ControllerHandler } from "./controller/controller";
+export { Route, RouteConfigurator, type RouteHandler } from "./route/route";
+export { ExpressAdapter } from "./express-adapter/adapter";
+export { type RouteConfiguration, type ArrowExpress } from "./types";
 export { RequestError } from "./error/request.error";